+++ spec/helpers/global.mjs
@@ -1,4 +1,4 @@
-import { parseUri, setSld } from "../../src/index.js";
+import { parseUri, setTlds } from "../../src/index.js";
 
-parseUri.setSld = setSld;
+parseUri.setTlds = setTlds;
 globalThis.parseUri = parseUri;